% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serror.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the Standard Error}
\usage{
standard_error(c(1,2,3,4,5))
}
\arguments{
\item{x a numeric vector}
}
\value{
the standard error of all the elements of the vector
}
\description{
Calculate the Standard Error
}
\examples{
standard_error(c(1,2,3,4,5))
}

