% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stddevr.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Gets the standard deviation of values in x.}
\usage{
standard_deviation(c(1,2,3,4,5))
}
\arguments{
\item{x a numeric vector}
}
\value{
the standard deviation of x
}
\description{
Gets the standard deviation of values in x.
}
\examples{
standard_deviation(c(1,2,3,4,5))
}
