% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serror.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the Standard Error}
\usage{
standard_error(x)
}
\arguments{
\item{x:}{a numeric vector}
}
\value{
a number (the standard error of all the elements of the vector)
}
\description{
Calculate the Standard Error
}
\examples{
standard_error(seq(1:100,2))
}

